Webmachine::Adapters::CryptosphereReel adapter
This is a reimplementation of Webmachine::Adapters::Reel which is mostly a
reworked version of the upstream code.

In addition to some refactoring, this adapter adds support for WebDAV verbs by
treating them like a POST request with a magical "_method" parameter. These
methods are used by git's smart HTTP protocol, so we need a way to support them
if we want to write a Webmachine-based git server.

The long-term goal of this adapter isn't for it to live inside of the
Cryptosphere forever, but to eventually get the changes back upstream into
Webmachine. In the meantime, this provides us a way to rapidly add features to
the adapter that the Cryptosphere actually needs without having to get all these
changes back upstream into Webmachine.